Large refactoring of all modules and functions to expose specific functions and modules at the top-level
